Merge in AILAB/replay from apply_in_pandas to master

* commit 'e554f309c1507d15ef53d880591c6bb65f858244':
  tests/models/test_pop_rec.py
  typo
  лишние print-ы
  scenatio test fix
  double  вместо float
  убрала warning Using or importing the ABCs from 'collections' instead of from 'collections.abc' is deprecated since Python 3.3,and in 3.9 it will stop working
  замена apply на applyInPandas
  applyInPandas with DDL string
  piu
  пример applyInPandas

+++ replay/data_preparator.py
@@ -10,7 +10,7 @@ from typing import Any, Dict, Iterable, List, Optional, Set, Tuple, Union
 
 from pyspark.sql import Column, DataFrame
 from pyspark.sql import functions as sf
-from pyspark.sql.types import FloatType, StringType, TimestampType
+from pyspark.sql.types import DoubleType, StringType, TimestampType
 
 from replay.constants import AnyDataFrame
 from replay.utils import convert2spark
@@ -387,7 +387,7 @@ class DataPreparator:
         }
         optional_columns = {
             "timestamp": ("1999-05-01", TimestampType()),
-            "relevance": (1.0, FloatType()),
+            "relevance": (1.0, DoubleType()),
         }
         if features_columns is None:
             features_columns = []
